Converted app and JavaScript to pass integer.
This is a fairly substantial rewrite of the code. Previously the
JavaScript would send the obtained prices to the Pebble as string
values. The app has now been rewritten so the JavaScript passes integer
values to the Pebble.

The idea behind this rewrite was to increase the speed at which the
Pebble code could parse incoming messages. It was my hope that making
the task faster would reduce the number of errors that stem from the
app being busy and therefore unable to accept new messages from the
JavaScript running on the phone.