Added support for displaying exchange volume.
Added support for transmitting and displaying the trade volume for an
exchange. Doing this required converting the volume to a byte array in
the JavaScript (as it often overflowed a 32-bit value), transmitting
the value to the Pebble, unpacking the value into a 64-bit variable,
and displaying it.

Fuck me.