Changed layout of the Exchange Detail screen.
The layout of the Exchange Detail screen has been changed to make it
more compact. I was getting annoyed at clicking the down button a
bajillion (technical measurement) times to see values.

The volume value has a propensity to overflow its text layer but that’s
fine. If the value is larger than the layer is gets truncated and “…”
is added to denote additional data exists.