Replaced malloc calls with calloc.
It turns out that the Pebble SDK, even though it’s not in the
documentation, supports calloc. That being the case I replaced the
malloc calls with calloc (even when it didn’t necessarily make senes
because I’m a big fan of known states even if it does cost a little
time in zeroing out memory).